requirements.txt
SQLAlchemy~=2.0.23
requests~=2.31.0
lxml~=4.9.3
Pillow~=10.0.1
python-telegram-bot
python-telegram-bot[job-queue]
